# feat: Added preSerialization hook and route option.