--- a/lib/hooks.js
+++ b/lib/hooks.js
@@ -4,6 +4,7 @@ const supportedHooks = [
   'onRequest',
   'preParsing',
   'preValidation',
+  'preSerialization',
   'preHandler',
   'onResponse',
   'onSend',
@@ -23,6 +24,7 @@ function Hooks () {
   this.onRequest = []
   this.preParsing = []
   this.preValidation = []
+  this.preSerialization = []
   this.preHandler = []
   this.onResponse = []
   this.onSend = []
@@ -47,6 +49,7 @@ function buildHooks (h) {
   hooks.onRequest = h.onRequest.slice()
   hooks.preParsing = h.preParsing.slice()
   hooks.preValidation = h.preValidation.slice()
+  hooks.preSerialization = h.preSerialization.slice()
   hooks.preHandler = h.preHandler.slice()
   hooks.onSend = h.onSend.slice()
   hooks.onResponse = h.onResponse.slice()