--- a/lib/reply.js
+++ b/lib/reply.js
@@ -129,8 +129,9 @@ Reply.prototype.send = function (payload) {
     if (hasContentType === false || contentType.indexOf('charset') === -1) {
       this[kReplyHeaders]['content-type'] = CONTENT_TYPE.JSON
     }
-    payload = serialize(this.context, payload, this.res.statusCode)
-    flatstr(payload)
+
+    preserializeHook(this, payload)
+    return
   }
 
   onSendHook(this, payload)
@@ -218,6 +219,32 @@ Reply.prototype.redirect = function (code, url) {
 
 Reply.prototype.callNotFound = function () {
   notFound(this)
+}
+
+function preserializeHook (reply, payload) {
+  if (reply.context.preSerialization !== null) {
+    onSendHookRunner(
+      reply.context.preSerialization,
+      reply.request,
+      reply,
+      payload,
+      preserializeHookEnd
+    )
+  } else {
+    preserializeHookEnd(null, reply.request, reply, payload)
+  }
+}
+
+function preserializeHookEnd (err, request, reply, payload) {
+  if (err != null) {
+    onErrorHook(reply, err)
+    return
+  }
+
+  payload = serialize(reply.context, payload, reply.res.statusCode)
+  flatstr(payload)
+
+  onSendHook(reply, payload)
 }
 
 function onSendHook (reply, payload) {