I've added preSerialization hook and route option. This fixes #1369.
One use-case for this is to being able to intercept serialization error before fast-json-stringify (which would hide all information since the error thrown has no context).

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [x] tests and/or benchmarks are included
- [x] documentation is changed or added
- [x] commit message and code follows [Code of conduct](https://github.com/fastify/fastify/blob/master/CODE_OF_CONDUCT.md)