@mcollina 2019-01-25
I wouldn’t expect this PR to be faster in any way. Can you confirm that?