@ShogunPanda 2019-01-25
I also updated typings, which I completedly forgot.