@ShogunPanda 2019-01-28
Types tests added.